+++ CHANGELOG.md
@@ -3,6 +3,16 @@
 All notable changes to this project will be documented in this file.
 This format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.2.1] - 2025-07-17
+Patch release with bug fixes and documentation updates.
+
+### Fixed
+ - Not casting batch to tuple in the `build_image_dataloader` function
+ - Missing value data imputation for tabular data
+ - Removed hardcoded `with_fit` and `with_onehot` parameters in the `TabularDataset` class
+ - Support for `Path` format in the `TabularDataset` class
+
+
 ## [0.2.0] - 2025-06-18
 Minor release with new Time Series metrics, bug fixes, and extended documentation.